# BE/students-swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Student Feedback API",
    "version": "1.0.1",
    "description": "학생 목록, 정오답, AI 피드백 및 최종 코멘트 관리 API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/students/all": {
      "get": {
        "summary": "학생 ID와 이름 목록 조회",
        "description": "DB에서 학생의 id와 이름만 조회합니다.",
        "responses": {
          "200": {
            "description": "학생 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "name": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "500": { "description": "DB 조회 실패" }
        }
      }
    },

    "/students/all/correct-wrong": {
      "get": {
        "summary": "모든 학생 정오답 조회",
        "responses": {
          "200": {
            "description": "학생 정오답 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "name": { "type": "string" },
                      "answers": {
                        "type": "array",
                        "items": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": { "description": "DB 조회 실패" }
        }
      }
    },

    "/students/{id}/correct-wrong": {
      "get": {
        "summary": "특정 학생의 정오답 정보 조회",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "학생 정오답 정보",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "name": { "type": "string" },
                    "answers": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "학생 정보를 찾을 수 없음" },
          "500": { "description": "DB 조회 실패" }
        }
      }
    },

    "/api/images/generate-all-students": {
      "post": {
        "summary": "모든 학생의 코멘트 이미지 일괄 생성",
        "description": "데이터베이스의 모든 학생에 대해 최종 코멘트가 포함된 이미지를 생성합니다.",
        "responses": {
          "200": {
            "description": "이미지 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "totalImages": { "type": "integer" },
                    "downloadUrl": { "type": "string" }
                  }
                }
              }
            }
          },
          "500": { "description": "이미지 생성 실패" }
        }
      }
    },

    "/students/all/ids": {
      "get": {
        "summary": "학생 ID와 이름 목록 조회 (이미지 생성용)",
        "description": "이미지 생성을 위한 학생 ID와 이름 목록을 조회합니다.",
        "responses": {
          "200": {
            "description": "학생 ID 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "name": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "500": { "description": "DB 조회 실패" }
        }
      }
    },

    "/api/images/generate-student/{id}": {
      "post": {
        "summary": "특정 학생의 코멘트 이미지 생성",
        "description": "지정된 학생의 최종 코멘트가 포함된 이미지를 생성합니다.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "학생 ID (UUID)"
          }
        ],
        "responses": {
          "200": {
            "description": "이미지 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "studentName": { "type": "string" },
                    "imageUrl": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "description": "학생을 찾을 수 없음" },
          "500": { "description": "이미지 생성 실패" }
        }
      }
    },

    "/comments/final/search": {
      "get": {
        "summary": "특정 학생 이름으로 코멘트 검색",
        "description": "학생 이름 일부 또는 전체를 이용해 해당 학생의 최종 코멘트를 검색합니다. 검색된 학생의 이름과 코멘트만 반환합니다.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "검색할 학생 이름 (부분 일치 가능)"
          }
        ],
        "responses": {
          "200": {
            "description": "검색된 학생의 최종 코멘트",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "name": { "type": "string" },
                    "comment": { "type": "string" }
                  },
                  "example": {
                    "success": true,
                    "name": "이하영",
                    "comment": "이번 단원에서는 문제 해결 과정이 명확하고 성실히 참여했습니다."
                  }
                }
              }
            }
          },
          "400": { "description": "잘못된 요청 (name 파라미터 누락 등)" },
          "404": { "description": "학생 또는 코멘트를 찾을 수 없음" },
          "500": { "description": "검색 중 서버 오류 발생" }
        }
      }
    },
  "/comments/final/all": {
    "get": {
      "summary": "전체 학생 최종 코멘트 조회 (3명씩 분할)",
      "description": "모든 학생의 최종 코멘트를 조회합니다. 응답이 길어지는 것을 방지하기 위해 3명씩 묶어서 반환합니다.",
      "responses": {
        "200": {
          "description": "전체 학생 코멘트 (3명씩 분할)",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "success": { "type": "boolean" },
                  "total": { "type": "integer", "description": "전체 학생 수" },
                  "chunks": { "type": "integer", "description": "분할된 청크 수" },
                  "data": {
                    "type": "array",
                    "description": "3명씩 묶인 학생 코멘트 배열",
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "comment": { "type": "string" }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "success": true,
                  "total": 5,
                  "chunks": 2,
                  "data": [
                    [
                      { "name": "박정구", "comment": "박정구 학생은..." },
                      { "name": "김준수", "comment": "김준수 학생은..." },
                      { "name": "이준우", "comment": "이준우 학생은..." }
                    ],
                    [
                      { "name": "카리나", "comment": "카리나 학생은..." },
                      { "name": "조세호", "comment": "조세호 학생은..." }
                    ]
                  ]
                }
              }
            }
          }
        },
        "404": { "description": "조회된 학생이 없음" },
        "500": { "description": "서버 오류" }
      }
    }
  },
    "/admin/delete-all": {
      "delete": {
        "summary": "DB 전체 삭제 (관리자용, 복구 불가)",
        "description": "StudentComment, StudentData, ClassRecord 데이터를 모두 삭제합니다. 요청 시 반드시 `?confirm=true` 쿼리 파라미터를 포함해야 실행됩니다.",
        "parameters": [
          {
            "name": "confirm",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "enum": ["true"] },
            "description": "삭제 실행 확인. 반드시 true로 설정해야 함"
          }
        ],
        "responses": {
          "200": {
            "description": "삭제 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "deleted": {
                      "type": "object",
                      "properties": {
                        "studentComments": { "type": "integer" },
                        "studentData": { "type": "integer" },
                        "classRecords": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "confirm 파라미터 누락 또는 false"
          },
          "500": {
            "description": "DB 삭제 중 서버 오류 발생",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "error": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}